fix: add dependency injected objects (for test)
