+++ pkg/timeline/service/account.test.ts
@@ -1,14 +1,8 @@
 import { Option, Result } from '@mikuroxina/mini-fn';
 import { describe, expect, it, vi } from 'vitest';
 
-import {
-  InMemoryAccountFollowRepository,
-  InMemoryAccountRepository,
-} from '../../accounts/adaptor/repository/dummy.js';
 import { Account, type AccountID } from '../../accounts/model/account.js';
-import { FetchService as AccountFetchService } from '../../accounts/service/fetch.js';
-import { FetchFollowService } from '../../accounts/service/fetchFollow.js';
-import { AccountModuleFacade } from '../../intermodule/account.js';
+import { dummyAccountModuleFacade } from '../../intermodule/account.js';
 import type { PartialAccount } from '../../intermodule/account.js';
 import { Note, type NoteID } from '../../notes/model/note.js';
 import { InMemoryTimelineRepository } from '../adaptor/repository/dummy.js';
@@ -16,13 +10,9 @@ import { AccountTimelineService } from './account.js';
 import { NoteVisibilityService } from './noteVisibility.js';
 
 describe('AccountTimelineService', () => {
-  const accountRepository = new InMemoryAccountRepository([]);
-  const accountFollowRepository = new InMemoryAccountFollowRepository();
-  const accountModule = new AccountModuleFacade(
-    new AccountFetchService(accountRepository),
-    new FetchFollowService(accountFollowRepository, accountRepository),
+  const noteVisibilityService = new NoteVisibilityService(
+    dummyAccountModuleFacade,
   );
-  const noteVisibilityService = new NoteVisibilityService(accountModule);
 
   const dummyPublicNote = Note.new({
     id: '1' as NoteID,
@@ -100,9 +90,11 @@ describe('AccountTimelineService', () => {
   });
 
   it('if following', async () => {
-    vi.spyOn(accountModule, 'fetchFollowers').mockImplementation(async () => {
-      return Result.ok([partialAccount1]);
-    });
+    vi.spyOn(dummyAccountModuleFacade, 'fetchFollowers').mockImplementation(
+      async () => {
+        return Result.ok([partialAccount1]);
+      },
+    );
     const res = await accountTimelineService.handle('100' as AccountID, {
       id: '101' as AccountID,
       hasAttachment: false,
@@ -118,9 +110,11 @@ describe('AccountTimelineService', () => {
   });
 
   it('if not following', async () => {
-    vi.spyOn(accountModule, 'fetchFollowers').mockImplementation(async () => {
-      return Result.ok([partialAccount1]);
-    });
+    vi.spyOn(dummyAccountModuleFacade, 'fetchFollowers').mockImplementation(
+      async () => {
+        return Result.ok([partialAccount1]);
+      },
+    );
     const res = await accountTimelineService.handle('100' as AccountID, {
       id: '0' as AccountID,
       hasAttachment: false,
